Move in-memory Join from unspilled to nonspilling sub-package
The `unspilled` package is too close to "unspilling, act of reading
spilled data back" not to be confusing as a package name. It remains
confusing even to those familiar with the code. `nonspilling`
sub-package is determined as a somewhat better name.
