+++ core/trino-main/src/main/java/io/trino/operator/join/nonspilling/JoinProbe.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.operator.join.unspilled;
+package io.trino.operator.join.nonspilling;
 
 import com.google.common.primitives.Ints;
 import io.trino.operator.join.LookupSource;
